repairgo-backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
email-validator
passlib
bcrypt
PyJWT
python-dotenv
numpy
pandas
scikit-learn
joblib
